+++ R/misc_and_utility.R
@@ -92,7 +92,7 @@ libs <- function(...) {
 #' @param df an [data.frame()] to select on
 #' @param rows Rows to keep (optional, incomplete rows still discarded)
 #' @param cols Cols to keep
-#' @param drop Whether to return a vecotr when only 1 column is selected (default TRUE)
+#' @param drop Whether to return a vector when only 1 column is selected (default TRUE)
 #' @return - Complete rows and (optionally) selected columns
 #' @export
 #' @family Data Functions
@@ -6075,6 +6075,7 @@ umx_long2wide <- function(data, famID = NA, twinID = NA, zygosity = NA, vars2kee
 #' @return - long-format dataframe
 #' @export
 #' @family Twin Data functions
+#' @seealso [reshape()]
 #' @examples
 #' long = umx_wide2long(data = twinData, sep = "")
 #' long = umx_wide2long(data = twinData, sep = "", verbose = TRUE)
@@ -7467,14 +7468,22 @@ umxHetCor <- function(data, ML = FALSE, use = c("pairwise.complete.obs", "comple
 			data[,i] = factor(data[,i])
 		}
 	}
-	hetc = hetcor(data, ML = ML, use = use, std.err = std.err)
-	if(verbose){
-		print(hetc)
-	}
-	if(return == "correlations"){
-		return(hetc$correlations)
+	if(is.null(dim(data)) || dim(data)[2]==1){
+		if(is.null(dim(data))){
+			return(var(as.numeric(data), na.rm = TRUE))
+		} else {
+			return(var(as.numeric(data[,1]), na.rm = TRUE))
+		}
 	} else {
-		return(hetc)
+		hetc = hetcor(data, ML = ML, use = use, std.err = std.err)
+		if(verbose){
+			print(hetc)
+		}
+		if(return == "correlations"){
+			return(hetc$correlations)
+		} else {
+			return(hetc)
+		}
 	}
 }